Solve lint conflicts
