--- clientScripts/socketList.js
@@ -1,2 +1,0 @@
-const socketList = { data: [] };
-export default socketList;
